Factor export() functions out of treemanager. This keeps the profiler and tool code from adding exponentially more code to the treemanager module.
